Fixed CreateOrder()  + CancelOrder() + ViewOrderDetail() v2

+++ customerApp/src/main/java/com/spring2020/customerapp/service/OrderService.java
@@ -9,7 +9,7 @@ import org.springframework.data.domain.Pageable;
 public interface OrderService {
     OrderDto createOrder(CreateOrderDto orderDto);
 
-    void cancelOrder(int id);
+    void cancelOrder(Long id);
 
     OrderDetailDto viewOrderDetail(Long id);
